+++ lib/drivers/provider/vmx/config.go
@@ -41,8 +41,8 @@ type Config struct {
 	//   for disk caching)
 	// * Positive (>0) value could also be available (but check it in your vmware dist in advance)
 	//   Please be careful here - noone wants the VM to fail allocation because of that...
-	CPUAlter int `json:"cpu_alter"` // 0 do nothing, <0 reduces number available CPUs, >0 increases it (dangerous)
-	RAMAlter int `json:"ram_alter"` // 0 do nothing, <0 reduces amount of available RAM (GB), >0 increases it (dangerous)
+	CPUAlter int32 `json:"cpu_alter"` // 0 do nothing, <0 reduces number available CPUs, >0 increases it (dangerous)
+	RAMAlter int32 `json:"ram_alter"` // 0 do nothing, <0 reduces amount of available RAM (GB), >0 increases it (dangerous)
 
 	// Overbook options allows tenants to reuse the resources
 	// It will be used only when overbook is allowed by the tenants. It works by just adding those
@@ -50,8 +50,8 @@ type Config struct {
 	// and want to run 2 tenants with requirement of 14 CPUs each - you can put 12 in CPUOverbook -
 	// to have virtually 28 CPUs. 3rd will not be running because 2 tenants will eat all 28 virtual
 	// CPUs. Same applies to the RamOverbook.
-	CPUOverbook uint `json:"cpu_overbook"` // How much CPUs could be reused by multiple tenants
-	RAMOverbook uint `json:"ram_overbook"` // How much RAM (GB) could be reused by multiple tenants
+	CPUOverbook uint32 `json:"cpu_overbook"` // How much CPUs could be reused by multiple tenants
+	RAMOverbook uint32 `json:"ram_overbook"` // How much RAM (GB) could be reused by multiple tenants
 
 	DownloadUser     string `json:"download_user"`     // The user will be used in download operations
 	DownloadPassword string `json:"download_password"` // The password will be used in download operations
@@ -118,7 +118,7 @@ func (c *Config) Validate() (err error) {
 		return err
 	}
 
-	if c.CPUAlter < 0 && cpuStat <= -c.CPUAlter {
+	if c.CPUAlter < 0 && cpuStat <= int(-c.CPUAlter) {
 		return log.Errorf("VMX: |CpuAlter| can't be more or equal the available Host CPUs: |%d| > %d", c.CPUAlter, cpuStat)
 	}
 
@@ -128,7 +128,7 @@ func (c *Config) Validate() (err error) {
 	}
 	ramStat := memStat.Total / 1073741824 // Getting GB from Bytes
 
-	if c.RAMAlter < 0 && int(ramStat) <= -c.RAMAlter {
+	if c.RAMAlter < 0 && ramStat <= uint64(-c.RAMAlter) {
 		return log.Errorf("VMX: |RamAlter| can't be more or equal the available Host RAM: |%d| > %d", c.RAMAlter, ramStat)
 	}
